WUFR-1165: Sensor board/22 support (#10)
* WUFR-1165: Sensor board 22 firmware support

Fixes WUFR-1166 WUFR-1168

* Add flashy status light and fix hsd bug

* WUFR-754: Make the ADC slow to increase precision

* Configure ADC pins as inputs

* Fuck ADC1, all my homies hate ADC1

For some reason, the presence of ADC1 being enabled in slave mode
would corrupt the reading on AIN0, the first sensor. This commit adds
a feature flag to disable dual-ADC mode entirely.

Now that I think about it, the reason for the issue is probably that
the second ADC wasn't configured to read anything, so when it was
triggered to start it would read one channel anyway. Since reverse_adc1
was set to the default value (0 == sensor1), it corrupted the reading
for sensor1.

So two solutions to this problem:
- disable this new feature flag when only using channels on ADC0
- jk this is the only reasonable solution

* Something is causing standard CAN frames to have ID 0

This is as it appears in my makeshift "serial_buffer" arduino sketch.
(https://confluence.washuracing.com/x/bgkm)
Don't know if zero IDs are being transmitted or if my receiver is dumb.

* Add further documentation to sensor/22